+++ CHANGES.rst
@@ -15,6 +15,8 @@ New Features:
   inline title* style), so it can be overridden (default: space character).
 * Warn about targets in ``rinoh_targets`` that are not defined in
   ``rinoh_documents``.
+* UpDownExpandingContainer: a container that symetrically expands both upwards
+  and downwards.
 
 Changed: